extract insights display widget
